decompose array compare logic, so that it can be used independently from json
